+++ lib/td_dd/grants/search.ex
@@ -35,10 +35,13 @@ defmodule TdDd.Grants.Search do
 
   def get_filter_values(user_or_permissions, params) do
     user_defined_filters = Query.create_filters(params, @index)
-    filter = case user_or_permissions do
-      [_h | _t] = permissions -> Query.create_filter_clause(permissions, user_defined_filters)
-      user_id -> Query.create_filter_clause_by(user_id, user_defined_filters)
-    end
+
+    filter =
+      case user_or_permissions do
+        [_h | _t] = permissions -> Query.create_filter_clause(permissions, user_defined_filters)
+        user_id -> Query.create_filter_clause_by(user_id, user_defined_filters)
+      end
+
     query = Query.create_query(%{}, filter)
     search = %{query: query, aggs: Query.get_aggregation_terms(@index)}
     Search.get_filters(search, @index)
@@ -71,6 +74,7 @@ defmodule TdDd.Grants.Search do
   end
 
   def search_by_user(params, user_name, page \\ 0, size \\ 50, index \\ :grants)
+
   def search_by_user(params, user_name, page, size, index) do
     filter(params, user_name, page, size, index)
   end
@@ -93,10 +97,13 @@ defmodule TdDd.Grants.Search do
 
   defp filter(params, user_or_permissions, page, size, index) do
     user_defined_filters = Query.create_filters(params, index)
-    filter = case user_or_permissions do
-      [_h | _t] = permissions -> Query.create_filter_clause(permissions, user_defined_filters)
-      user_id -> Query.create_filter_clause_by(user_id, user_defined_filters)
-    end
+
+    filter =
+      case user_or_permissions do
+        [_h | _t] = permissions -> Query.create_filter_clause(permissions, user_defined_filters)
+        user_id -> Query.create_filter_clause_by(user_id, user_defined_filters)
+      end
+
     query = Query.create_query(params, filter)
     sort = Map.get(params, "sort", default_sort(index))