+++ lib/td_dd_web/views/domain_view.ex
@@ -1,0 +1,7 @@
+defmodule TdDdWeb.DomainView do
+  use TdDdWeb, :view
+
+  def render("embedded.json", %{domain: domain}) do
+    Map.take(domain, [:id, :external_id, :name])
+  end
+end