+++ test/support/cache_helpers.ex
@@ -5,8 +5,11 @@ defmodule CacheHelpers do
 
   import TdDd.Factory
 
+  alias TdCache.AclCache
   alias TdCache.ConceptCache
   alias TdCache.LinkCache
+  alias TdCache.Permissions
+  alias TdCache.Redix
   alias TdCache.TaxonomyCache
   alias TdCache.TemplateCache
   alias TdCache.UserCache
@@ -14,7 +17,7 @@ defmodule CacheHelpers do
 
   def insert_domain(params \\ %{}) do
     parent_ids = Map.get(params, :parent_ids)
-    %{id: domain_id} = domain = build(:domain)
+    %{id: domain_id} = domain = build(:domain, params)
     domain = if is_nil(parent_ids), do: domain, else: Map.put(domain, :parent_ids, parent_ids)
     TaxonomyCache.put_domain(domain)
     ExUnit.Callbacks.on_exit(fn -> TaxonomyCache.delete_domain(domain_id) end)
@@ -83,5 +86,28 @@ defmodule CacheHelpers do
     {:ok, _} = UserCache.put(user)
     ExUnit.Callbacks.on_exit(fn -> UserCache.delete(id) end)
     user
+  end
+
+  def insert_acl(domain_id, role, user_ids) do
+    AclCache.set_acl_role_users("domain", domain_id, role, user_ids)
+
+    ExUnit.Callbacks.on_exit(fn ->
+      AclCache.delete_acl_roles("domain", domain_id)
+      AclCache.delete_acl_role_users("domain", domain_id, role)
+    end)
+
+    :ok
+  end
+
+  def insert_grant_request_approver(user_id, domain_id, role_name \\ "grant_approver") do
+    ExUnit.Callbacks.on_exit(fn ->
+      UserCache.delete(user_id)
+      Redix.command!(["SREM", "permission:approve_grant_request:roles", role_name])
+    end)
+
+    insert_user(id: user_id)
+    insert_acl(domain_id, role_name, [user_id])
+    UserCache.put_roles(user_id, %{role_name => [domain_id]})
+    Permissions.put_permission_roles(%{"approve_grant_request" => [role_name]})
   end
 end