+++ test/td_dd/search/structure_enricher_test.exs
@@ -6,14 +6,15 @@ defmodule TdDd.Search.StructureEnricherTest do
   @moduletag sandbox: :shared
 
   setup do
-    domain = CacheHelpers.insert_domain()
+    %{id: parent_id} = parent_domain = CacheHelpers.insert_domain()
+    domain = CacheHelpers.insert_domain(%{parent_ids: [parent_id]})
     %{id: template_id, name: template_name} = template = CacheHelpers.insert_template()
 
     CacheHelpers.insert_structure_type(name: template_name, template_id: template_id)
 
     # insert template and structure type
     start_supervised!(StructureEnricher)
-    [domain: domain, template: template]
+    [domain: domain, template: template, parent_domain: parent_domain]
   end
 
   describe "StructureEnricher.refresh/1" do
@@ -30,6 +31,21 @@ defmodule TdDd.Search.StructureEnricherTest do
                |> StructureEnricher.enrich()
 
       assert %{id: ^domain_id, name: ^domain_name, external_id: ^domain_external_id} = domain
+    end
+
+    test "enriches the domain parents", %{
+      domain: %{id: domain_id, name: domain_name, external_id: domain_external_id},
+      parent_domain: %{id: parent_id, name: parent_name, external_id: parent_external_id}
+    } do
+      assert %{domain_parents: parents} =
+               :data_structure
+               |> insert(domain_id: domain_id)
+               |> StructureEnricher.enrich()
+
+      assert parents == [
+               %{id: domain_id, external_id: domain_external_id, name: domain_name},
+               %{id: parent_id, external_id: parent_external_id, name: parent_name}
+             ]
     end
 
     test "enriches the link count" do