+++ test/td_dq_web/controllers/rule_filter_controller_test.exs
@@ -2,10 +2,6 @@ defmodule TdDqWeb.RuleFilterControllerTest do
   use TdDqWeb.ConnCase
   use PhoenixSwagger.SchemaTest, "priv/static/swagger_dq.json"
 
-  setup %{conn: conn} do
-    {:ok, conn: put_req_header(conn, "accept", "application/json")}
-  end
-
   describe "index" do
     @tag authentication: [role: "admin"]
     test "search filters should return at least the informed filters", %{conn: conn} do